fix: Fix parameters in GetUserById, merging conflict

+++ src/Modules/Tours/Explorer.Tours.Core/UseCases/TourService.cs
@@ -32,7 +32,7 @@ namespace Explorer.Tours.Core.UseCases
         public Result<List<TourDto>> GetToursListByAuthor(long authorId, int page, int pageSize)
         {
 
-            var userResult = _userAccountService.GetUserById(authorId);
+            var userResult = _userAccountService.GetUserById((int)authorId);
 
             if (userResult.IsSuccess && userResult.Value != null)
             {
